wado-lsp: move definition tests to tests/ integration harness
Relocate the 41 go-to-definition tests from the inline `#[cfg(test)]`
module in `src/definition.rs` into a dedicated integration test file at
`tests/definition.rs`. The tests now exercise the public `Engine::definition`
API instead of the private `find_definition` helper, which keeps the
public surface honest and trims ~775 lines from the source module.
